+++ src/helper/GlobalColors.cpp
